@@ -5,19 +5,19 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.18.RELEASE</version>
+        <version>3.0.2</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.a6raywa1cher</groupId>
     <artifactId>reschedule-tsu-vk</artifactId>
-    <version>1.2.5</version>
+    <version>1.2.6</version>
     <packaging>jar</packaging>
     <name>reschedule-tsu-vk</name>
     <description>Vk.com bot for TvSU</description>
 
     <properties>
-        <java.version>11</java.version>
-        <jackson.version>2.13.4</jackson.version>
+        <java.version>17</java.version>
+        <jackson.version>2.14.2</jackson.version>
         <maven.test.skip>true</maven.test.skip>
     </properties>
 
@@ -37,7 +37,7 @@
         <dependency>
             <groupId>com.vk.api</groupId>
             <artifactId>sdk</artifactId>
-            <version>1.0.6</version>
+            <version>1.0.10</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -110,7 +110,7 @@
         <dependency>
             <groupId>com.google.cloud</groupId>
             <artifactId>google-cloud-dialogflow</artifactId>
-            <version>0.110.0-alpha</version>
+            <version>4.16.0</version>
         </dependency>
 
         <dependency>
